feat: 회사 정보 폼 컴포넌트 분리
